requirements.txt
opencv-python
numpy
python-dotenv
scikit-learn
